+++ go/internal/latex/assemble.go
@@ -149,7 +149,7 @@ func (r *Runner) bookSessionTools(proj, buildDir string, compile *CompileTexTool
 		&WorkBashTool{Root: buildDir, Mounts: r.sessionMounts(kindBook, buildDir), TmpDir: bashTmp, MaxOutput: r.cfg.BashMaxOutput(), Sandbox: r.cfg.BashSandboxEnabled(), Python: r.pythonEnv(proj), Log: r.log, Tid: 1},
 		compile,
 		&ViewPDFTool{Mounts: r.sessionMounts(kindBook, buildDir), Comp: r.comp, SoftMax: r.cfg.ViewPDFMax(), WarnRatio: r.cfg.ViewWarnRatio()},
-		&ViewImageTool{Root: buildDir, SoftMax: r.cfg.ViewImageMax(), WarnRatio: r.cfg.ViewWarnRatio()},
+		&ViewImageTool{Root: buildDir, BareSearch: true, SoftMax: r.cfg.ViewImageMax(), WarnRatio: r.cfg.ViewWarnRatio()},
 		&ListFontsTool{FontsDir: r.cfg.Paths.Fonts},
 		submit,
 	}